--- src/main/java/Task8/Movable.java
@@ -1,13 +1,0 @@
-package Task8;
-
-public interface Movable {
-
-     void moveUp();
-
-     void moveDown();
-
-     void moveLeft();
-
-     void moveRight();
-
-}
