--- src/main/java/Task8/MovablePoint.java
@@ -1,54 +1,0 @@
-package Task8;
-
-public class MovablePoint implements Movable {
-    private int x;
-    private int y;
-    private int xSpeed;
-    private int ySpeed;
-
-
-
-    public MovablePoint(int x, int y) {
-        this.x = x;
-        this.y = y;
-    }
-
-    public MovablePoint(int x, int y, int xSpeed, int ySpeed) {
-        this.x = x;
-        this.y = y;
-        this.xSpeed = xSpeed;
-        this.ySpeed = ySpeed;
-    }
-
-    @Override
-    public void moveUp() {
-        y +=ySpeed;
-
-    }
-
-    @Override
-    public void moveDown() {
-        y -= ySpeed;
-    }
-
-    @Override
-    public void moveLeft() {
-        x -= xSpeed;
-
-    }
-
-    @Override
-    public void moveRight() {
-        x += xSpeed;
-    }
-
-    @Override
-    public String toString() {
-        return "MovablePoint{" +
-                "x=" + x +
-                ", y=" + y +
-                ", xSpeed=" + xSpeed +
-                ", ySpeed=" + ySpeed +
-                '}';
-    }
-}
