+++ src/main/java/task1/Main.java
@@ -1,0 +1,18 @@
+package task1;
+
+import java.util.Scanner;
+
+class Program {
+    public static void main(String[ ] args) {
+        Scanner sc = new Scanner(System.in);
+        int x = sc.nextInt();
+        int y = sc.nextInt();
+
+        Square a = new Square(x);
+        Circle b = new Circle(y);
+        a.area();
+        b.area();
+    }
+}
+
+
