+++ src/main/java/task4/Point2D.java
@@ -1,0 +1,42 @@
+package task4;
+
+public class Point2D {
+    public float x;
+    public float y;
+
+    public Point2D(float x, float y) {
+        this.x = x;
+        this.y = y;
+    }
+
+    public Point2D() {
+    }
+
+    public float getX() {
+        return x;
+    }
+
+    public float getY() {
+        return y;
+    }
+
+    public void setX(float x) {
+        this.x = x;
+    }
+
+    public void setY(float y) {
+        this.y = y;
+    }
+    public float[] getXY(){
+        return new float[]{x,y};
+    }
+    public void setXY (float x, float y){
+        this.x = x;
+        this.y = y;
+    }
+
+    @Override
+    public String toString() {
+        return String.format("(%f, %f)", x, y);
+    }
+}
