+++ src/main/java/task5/Staff.java
@@ -1,0 +1,39 @@
+package task5;
+
+
+public class Staff extends Person{
+    private String specialization;
+    private double salary;
+
+    public Staff(String name, String address, String specialization, double salary) {
+        super(name, address);
+        this.specialization = specialization;
+        this.salary = salary;
+    }
+
+    public String getSpecialization() {
+        return specialization;
+    }
+
+    public void setSpecialization(String specialization) {
+        this.specialization = specialization;
+    }
+
+    public double getSalary() {
+        return salary;
+    }
+
+    public void setSalary(double salary) {
+        this.salary = salary;
+    }
+
+    @Override
+    public String toString() {
+        return "Staff{" +
+                "name='" + name + '\'' +
+                ", address='" + address + '\'' +
+                ", specialization='" + specialization + '\'' +
+                ", salary=" + salary +
+                '}';
+    }
+}
