+++ src/main/java/task8/Movable.java
@@ -1,0 +1,13 @@
+package task8;
+
+public interface Movable {
+
+     void moveUp();
+
+     void moveDown();
+
+     void moveLeft();
+
+     void moveRight();
+
+}
