+++ src/main/java/task9/GeometricObject.java
@@ -1,0 +1,8 @@
+package task9;
+
+public interface GeometricObject {
+
+    double getPerimeter();
+
+    double getArea();
+}
