+++ src/main/java/task9/Resizable.java
@@ -1,0 +1,5 @@
+package task9;
+
+public interface Resizable {
+    void resizable (int percent);
+}
