+++ src/main/java/task9/ResizableCircle.java
@@ -1,0 +1,21 @@
+package task9;
+
+public class ResizableCircle extends Circle implements Resizable{
+
+    public ResizableCircle(double radius) {
+        super(radius);
+    }
+
+    @Override
+    public void resizable(int percent) {
+        radius = radius * percent/ 100;
+
+    }
+
+    @Override
+    public String toString() {
+        return super.toString();
+    }
+
+
+}
